시간 타입 변경

+++ delivery/src/main/java/com/hubEleven/delivery/infrastructure/dto/HubRouteSegmentResponseDto.java
@@ -8,6 +8,6 @@ public record HubRouteSegmentResponseDto(
 		UUID toHubId,
 		Integer sequence,
 		Double distance,
-		Long duration) {
+		Integer duration) {
 
 }
