+++ delivery/src/main/java/com/hubEleven/deliveryRoute/application/dto/DeliveryRouteRequestDto.java
@@ -9,6 +9,6 @@ public record DeliveryRouteRequestDto(
 		Long deliveryManagerId,
 		DeliveryStatus status,
 		Double expectedDistance,
-		Long expectedDuration,
+		Integer expectedDuration,
 		Double actualDistance,
-		Long actualDuration) {}
+		Integer actualDuration) {}
